pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Andrey_Fominykh <meriadok2003@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
gunicorn = "^20.1.0"
django-heroku = "^0.3.1"
psycopg2 = "^2.9.3"
whitenoise = "^6.0.0"
dj-database-url = "^0.5.0"
requests = "^2.27.1"
django-bootstrap4 = "^22.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.5"
flake8 = "^4.0.1"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
